code/central/camera/camera_aliot/requirements.txt
aliot-py==1.0.5
gpiozero==2.0.1
picamera==1.13.0
